отладил функционал удаления поста
